LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/script-charset-02-expected.txt


FAIL Script @type: unknown parameters assert_not_equals: got disallowed value 5
FAIL Script @type: unknown parameters 1 assert_equals: expected 5 but got 10
